+++ mindone/trainers/zero.py
@@ -13,6 +13,234 @@ from .train_step import TrainOneStepWrapper
 _logger = logging.getLogger(__name__)
 
 
+hyper_map = ops.HyperMap()
+
+_optim_allgather = ops.MultitypeFuncGraph("optim_allgather")
+
+
+@_optim_allgather.register("Function", "Bool", "Tensor", "Tensor", "Bool")
+def _run_optim_allgather(allgather, last_assign, variable, value, need_allgather):
+    if need_allgather:
+        value = allgather(value)
+    if last_assign:
+        ops.assign(variable, value)
+    return True
+
+
+_dp_allreduce = ops.MultitypeFuncGraph("dp_allreduce")
+
+
+@_dp_allreduce.register("Function", "Tensor", "Tensor")
+def _run_dp_allreduce(dp_allreduce, dp_group_size, gradient):
+    gradient = dp_allreduce(gradient) / dp_group_size
+    return gradient
+
+
+_stage2_reduce_scatter = ops.MultitypeFuncGraph("stage2_reduce_scatter")
+
+
+@_stage2_reduce_scatter.register("Function", "Tensor", "Tensor", "Bool")
+def _run_stage2_reduce_scatter(reduce_scatter, op_group_size, gradient, need_reduce_scatter):
+    if need_reduce_scatter:
+        gradient = reduce_scatter(gradient) / op_group_size
+    return gradient
+
+
+_stage1_split_grad = ops.MultitypeFuncGraph("stage1_split_grad")
+
+
+@_stage1_split_grad.register("Function", "Int", "Tensor", "Bool")
+def _run_stage1_split_grad(split, op_rank_id, gradient, need_split):
+    if need_split:
+        gradient = split(gradient)[op_rank_id]
+    return gradient
+
+
+@ms.ms_class
+class ZeroHelper:
+    """
+    Zero redundancy optimizer(ZeRO) build helper.
+
+    - zero_stage is 0: Normal optimizer update.
+    - zero_stage is 1: Split optimizer parameters and gradients, manually updating optimizer parameters.
+    - zero_stage is 2: Split optimizer parameters, replace gradients allreduce with reducescatter,
+        manually updating optimizer parameters.
+    - zero_stage is 3: Split optimizer parameters, normal optimizer update.
+
+    Args:
+        optimizer (`nn.Optimizer`): Must be the subclass of MindSpore Optimizer.
+        zero_stage (`int`, *optional*): Stage setting of ZeRO, default is 0.
+        op_group (`str`, *optional*): The name of the optimizer parallel communication group, default is None.
+        dp_group (`str`, *optional*): The name of the data parallel communication group, default is None.
+        optimizer_offload (`bool`, *optional*): Only take effect when optimizer is AdamWeightDecay, default is False.
+    """
+
+    def __init__(
+        self,
+        optimizer: nn.Optimizer,
+        zero_stage: int = 0,
+        op_group: str = None,
+        dp_group: str = None,
+        optimizer_offload: bool = False,
+    ):
+        self.optimizer = optimizer
+        self.zero_stage = zero_stage
+        self.op_group = op_group
+        self.ori_parameters = self.optimizer._parameters
+        # Init parallel settings
+        self.is_parallel = _get_parallel_mode() == ParallelMode.DATA_PARALLEL
+        if self.is_parallel and self.zero_stage != 0:
+            _logger.warning("Not in DATA_PARALLEL, set zero_stage to 0.")
+            self.zero_stage = 0
+        self.split_op = ops.Identity()
+        self.op_allgather = ops.Identity()
+        self.op_reduce_scatter = ops.Identity()
+        self.dp_allreduce = ops.Identity()
+        self.op_group_size = get_group_size(self.op_group) if self.is_parallel else 1
+        self.op_rank_id = get_rank(self.op_group) if self.is_parallel else 0
+        self.need_dp = False
+        self.last_assign = False
+        self.dp_group_size = 1
+        self.need_allgather = [False] * len(self.optimizer._parameters)
+        if self.zero_stage in [1, 2] and self.is_parallel:
+            _logger.info("Clone optimizer.parameters, will increase memory.")
+            # Because the first input of MindSpore optimizer must be ms.Parameter,
+            # copy optimizer.parameters for optimizer parameters update.
+            # It will increase 1/n parameters' memory.
+            self.optimizer.parameters = self.optimizer.parameters.clone(prefix="wrapper", init="same")
+            self.optimizer._parameters = self.optimizer.parameters
+            self.last_assign = True
+        if self.zero_stage in [1, 2, 3] and self.is_parallel:
+            if self.zero_stage == 2:
+                self.op_reduce_scatter = ops.ReduceScatter(op=ops.ReduceOp.SUM, group=self.op_group)
+            if self.zero_stage in [1, 2]:
+                # AllGather the parameters after optimizer calculate to update the parameters in train network.
+                self.op_allgather = ops.AllGather(group=self.op_group)
+            self.need_dp = dp_group is not None
+            if self.need_dp:
+                # Set it when op_group is not the WORLD_COMM_GROUP.
+                self.dp_allreduce = ops.AllReduce(op=ops.ReduceOp.SUM, group=dp_group)
+                self.dp_group_size = ms.Tensor(get_group_size(group=dp_group), ms.float32)
+            self.split_op = ops.Split(0, self.op_group_size)  # optimizer parallel split
+            self.split_params()
+        self.need_allgather = tuple(self.need_allgather)
+        self.hyper_map = ops.HyperMap()
+        if optimizer_offload:
+            if isinstance(self.optimizer, nn.AdamWeightDecay):
+                nn.AdamWeightDecay.target("CPU")
+                _logger.info("Set optimizer run offload.")
+            else:
+                _logger.warning("optimizer_offload only take effect when optimizer is AdamWeightDecay.")
+                optimizer_offload = False
+        _logger.info(
+            f"Build TrainOneStepWrapper with ZeRO stage: {self.zero_stage}, "
+            f"optimizer_offload: {optimizer_offload}, "
+            f"op_group_size: {self.op_group_size} "
+            f"op_rank_id: {self.op_rank_id} "
+            f"dp_group_size: {self.dp_group_size} "
+        )
+
+    def split_param(self, param):
+        return self.split_op(param)[self.op_rank_id]
+
+    def get_optimizer_param_tuples(self):
+        param_tuples = []
+        if ms.get_context("mode") == ms.PYNATIVE_MODE:
+            for name in self.optimizer._params_list:
+                if name in ["_parameters", "parameters"]:
+                    continue
+                _logger.debug(f"Add optimizer param_tuples {name}")
+                param_tuples.append(getattr(self.optimizer, name))
+        else:
+            for attr in self.optimizer.__dict__:
+                if isinstance(getattr(self.optimizer, attr), ms.ParameterTuple):
+                    if attr in ["_parameters", "parameters"]:
+                        continue
+                    _logger.debug(f"Add optimizer param_tuples {attr}")
+                    param_tuples.append(getattr(self.optimizer, attr))
+        return param_tuples
+
+    def split_params(self):
+        param_tuples = self.get_optimizer_param_tuples()
+        for i, param in enumerate(self.optimizer._parameters):
+            _logger.debug(f"Split optimizer param {param.name} {param.shape}")
+            # If zero_stage is 3, the parameters in train network have been split,
+            # use parameter in param_tuples to get batch size.
+            if self.zero_stage == 3:
+                if param_tuples:
+                    B = param_tuples[0][i].shape[0]
+                else:
+                    continue
+            else:
+                B = param.shape[0]
+            _logger.debug(f"Do split with zero_stage {self.zero_stage}")
+            if param.parallel_optimizer and B >= self.op_group_size and B % self.op_group_size == 0:
+                if self.zero_stage in [1, 2]:
+                    self.need_allgather[i] = True
+                    ori_shape = param.shape
+                    param.assign_value(self.split_param(param))
+                    _logger.debug(f"Optimizer {param.name} from {ori_shape} to {param.shape}")
+                for param_tuple in param_tuples:
+                    ori_shape = param_tuple[i].shape
+                    param_tuple[i].assign_value(self.split_param(param_tuple[i]))
+                    _logger.debug(f"Optimizer {param_tuple[i].name} " f"from {ori_shape} to {param_tuple[i].shape}")
+
+    @ms.jit
+    def reduce_scatter_gradients(self, gradients):
+        dtype = gradients[0].dtype
+        gradients = self.hyper_map(
+            ops.partial(
+                _stage2_reduce_scatter,
+                self.op_reduce_scatter,
+                ms.Tensor(self.op_group_size, dtype),
+            ),
+            gradients,
+            self.need_allgather,
+        )
+        return gradients
+
+    @ms.jit
+    def dp_allreduce_gradients(self, gradients):
+        dtype = gradients[0].dtype
+        gradients = self.hyper_map(
+            ops.partial(
+                _dp_allreduce,
+                self.dp_allreduce,
+                ms.Tensor(self.dp_group_size, dtype),
+            ),
+            gradients,
+        )
+        return gradients
+
+    @ms.jit
+    def split_gradients(self, gradients):
+        gradients = self.hyper_map(
+            ops.partial(
+                _stage1_split_grad,
+                self.split_op,
+                self.op_rank_id,
+            ),
+            gradients,
+            self.need_allgather,
+        )
+        return gradients
+
+    @ms.jit
+    def run_optimizer(self, grads):
+        optim_result = self.optimizer(grads)
+        if self.zero_stage == 1 or self.zero_stage == 2:
+            optim_result = ops.depend(
+                self.hyper_map(
+                    ops.partial(_optim_allgather, self.op_allgather, self.last_assign),
+                    self.ori_parameters,
+                    self.optimizer._parameters,
+                    self.need_allgather,
+                ),
+                optim_result,
+            )
+        return optim_result
+
+
 class ZeroParamWrapper(nn.Cell):
     """
     a cell to Insert communication operators before and after parameters when `zero_stage == 3`.
@@ -22,7 +250,7 @@ class ZeroParamWrapper(nn.Cell):
         super().__init__(auto_prefix=False)
         self.op_group = op_group
         self.zero_stage = zero_stage
-        if zero_stage not in [2, 3]:
+        if zero_stage != 3:
             raise ValueError(f"ZeroParamWrapper not support zero_stage {zero_stage}.")
         self.need_rewrite = self.check_rewrite(param)
         # Init parallel settings
@@ -188,6 +416,7 @@ def prepare_train_network(
         raise ValueError("op_group {op_group} and dp_group {dp_group} not full network hccl group coverage")
 
     new_network = prepare_network(network, zero_stage, op_group)
+    zero_helper = ZeroHelper(optimizer, zero_stage, op_group, dp_group, optimizer_offload)
     train_network = TrainOneStepWrapper(
         new_network,
         optimizer,
@@ -199,9 +428,6 @@ def prepare_train_network(
         clip_grad=clip_grad,
         clip_norm=clip_norm,
         verbose=verbose,
-        zero_stage=zero_stage,
-        optimizer_offload=optimizer_offload,
-        op_group=op_group,
-        dp_group=dp_group,
+        zero_helper=zero_helper,
     )
     return train_network
